Split session implementation into separate files.
